SmartEating/backend/SmartEating/requirements.txt
asgiref==3.6.0
Django==3.2
Pillow==9.3.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.3
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
PyJWT==1.7.1
django-cors-headers